requirements-dev.txt
pytest~=6.2.4
hypothesis~=6.14.2
hypothesis_jsonschema~=0.20.1
jsonschema2md~=0.2.0.post1
PyYAML~=5.4.1
pre-commit==2.13.0
#
# Please keep the following reqs in sync /w .pre-commit-config.yaml
#
flake8==3.9.2
pylint==2.9.3
black==21.6b0
yamllint==1.26.1
